Add files via upload
The code contains the wifi manager to connect an existing wifi via user's input. After connecting to the wifi, it fetches the credentials from the API and test if one of those credentials are existing, if none of those connections are available, it will go back to the wifi manager. As for the rfid, the countdown starts when the card is connected to the reader. Even the card is removed during the countdown, it still resumes regardless the state of the reader (card is in the reader or not). If the countdown expires and card is in the reader, the card will be written off and replace with a dummy data. If the case that the countdown expires and the card was removed, if the card is placed in the reader, the   data will be also erased.